spikexplorer-core/pyproject.toml
[tool.poetry]
name = "spikexplorer-core"
version = "0.1.0"
description = ""
authors = ["Nafiul Nipu <nnipu2@uic.edu>"]
readme = "README.md"
packages = [{include = "spikexplorer_core"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
pandas = "^2.0.1"
scipy = "^1.10.1"
flask = "^2.3.2"
flask-cors = "^3.0.10"
pyarrow = "^11.0.0"
python-dotenv = "^1.0.0"
fastparquet = "^2023.4.0"
scikit-learn = "^1.2.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"
black = "^23.3.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
